feat(web): Refine color system
Refactor how variables are defined so that they are defined in css as
the actual hsl  deffinitions and not just the values. this gives
better support for ideas to do collor highlighting.

Refine the colors for succes wanr and destructive and skip the border
on them completely. These values are now added to figma as well.

Lets move on to more fun things!