feat(web): Vehicle: Add all fields to form + DataForm improvements
Important refactor:

Change the internal mechanics of DataForm.

I find the approach and design thinking the useEffects together with
useRefs to see if data or mode has just changed to problematic.

Pondering more on the nature of react fiber concludes that we should
derive our data for when we need direct data that cannot wait for
useEffect and build the useEffects so that they truly can take time
without causing bad UX.

The useEffect implementations in Field for form validation to be
toggled visiable is where it truly broke down. Here treat the body as
if it always comes in sync with mode and it does not. Mode causes
sideeffects that changes body. Instead we could solve this feature by
deriving data directly on render with useMemo inside DataForm.

Cool.