feat(web): Upgrade search bar and configure it for shop page
Refactor dropdownOptions to now be enums and structure the export so
that each enum list (prev options) represent a mirror of the object it
maps. This is needed for enum match fn to work (see next paragraph).

Now the search bar takes a match function instead of its built in
fuzzy finder. This function leaves the responsibility for each page to
define its own searching mechanism. The matchFn for the search bar for
shop includes three parts:

- enum match: use enumMatchFn from the new enums file and input the
  enum map into it. This allow searching strictly for enums and their
  localissed value.
- fuzzy find: allow the old fuzzy find algorithm, in the case of the
  shop page it is only used for comment.
- custom ticket match: Here we define our own logic for matching the
  search against the customers first name of the ticket in a fuzzy manner.