fix(web): Portal layout broken flex styles
Fix horisontal and vertical scroll and resolve so that table and props
panel grow correct but not beyond the viewport.

Unfortunately the vertical scrollbar in the table starts from the
hedaers instead of sitting next to the body only. The fix is to break
the header and body appart and synx the width of them using
javascript. That sound like a tough one... Lets not think about that now...