refactor(web): Feature based structure for files concering data
There is a lot of files that concern data for a feature:

* api requests
* form validation
* types (previously in the api file)
* mutations (hooks)
* queries (hooks)
* enums

I've always wanted to use feature based structuring but always found
that splitting everying by features loses some of the valuable
decoupling in a react app. This time around though it feels like a
good balance has been struck. React component tree logic reciced
together in root/ outside of the feature based structuring while
everything data related is feature structured.

This greatly means that we have a seperation of concern that also
bundles the data concern together so that when a change in backend is
done to some endpoints data, that developer only needs to make changes
in one directory...