feat: Tickets page in Portal
+ Add ticket page
  + (adds calendar component from shadcn)
  + DataForm.Select now has opt in multi select logic
  + DataForm.Switch for boolean values
  + DataForm.DatePicker for picking dates
  + DataTable.MultiSelect now can filter for array values on a row (ie
  locationIds, vehicleIds etc...)
  + Add employee api and hooks

* Tweak colors
* Inventory page alsu use useColumns instead

Comments about the work:

The tickets page only aggregates locationIds into the data passed to
the table, even though customers, employees and vehicles are also used
to render the table. This data is instead aggregated when creating the
column definitions for the table, doing this in a hook instead, called
useColumns().

Previsouly for the vehcicle page all data used inte the column
definitions was aggregated in the query hook creating the data passed
to the table.

While i like to aggregate all the data in one place and abstract the
column definitions away from such logic now that we have
useXxxAsEnumsQ queries consuming all the data from other endpoints
directly as enums makes the aggregation in columns quite sleek.

The reason for using the enums hooks in useColumns and not when
aggregating the data is to keep the data flat and minimal and not
injecting all the enums with their metadata and icon components.

However locationIds needs to aggregated in the data query hook as the
filter function for filtering on locations filters on the the dataKey
on the data row. One could opt in here and write your own filter
function but a balance was struck.

Not sure which design pattern would be the best here, to be further evaulated...

Co-authored-by: Christopher Lo-Martire <christopher.lomartire@alten.se>