feat: SSG with preloaded data (from web-edit)
--

It is time to fetch the webedit data during build time so that it can
be preloaded in the react app.

We do this by embedding the data from webedit in a script tag
assigning it to a json string on the window object. A context is used
and populated by the data from renderSSR on server side and from
window.__PRELOADED_DATA__ on client side.

The data is fetched using the getDataForPreloadingServerSide exported
from the react app. This way we can use our normal axios instance and
webedit configure api to get the data.

This way the data is always available in the react app, looks
synchronous. Wow!

--

Routes where broken into ssr routes and csr routes since they have key
differences in their logic applied to them. instead of keeping them
together. However they are kind of double edged, the CSR routes are
not localized and SSR routes are the localised ones. They might as
well be called something else but they also happen to be CSR and SSR
respectively....

--

The dev server also had to be updated to reflect the new build logic.

--

utils in now a bit bloated and it would be good to break up into a
directory of different utils.

--

While the useStrictContext is nice i find a bit unnessecary, clear
dumb code trumps smart code in most cases. In this case i wanted to
have some fun and try it out and could not bear to remove it. Lets go
with it