fix(web): Bad practice for clearing all filters
The current filters, i.e. <FilterMultiSelect /> and <FilterSearch /> store their own state as to weather or not they have any active filtering, and they have a useEffect to update the filterFns through the useDataTable() hook through a master/slave relationship. The filters are the source of truth, i.e. master, and the data table are slaves to that.

However there is an issue here. <FilterClear /> that can clear active filters does not go to the source of thruth to say "it is time to reset your state", instead it removes all the registered filterFn in the data table and the filters have a useEffect that clears its state if there are no filterFns. We have a middleman...

This has worked so far but actually prevents certain use cases. Also it is a bad practice. Now the source of thruth is split. Sometimes the filters are the master and sometimes the data table. The <FilterClear /> button should ask the filters directly to clear their state and the filterFns should be updated accordingly as the filters are masters.

The issue that brought up this bad implementation is the fact that the state in a <FilterMultiSelect /> cannot be set with an initial state. If an initial state is passed to its useState() hook, since there are no filterFns yet, the useEffect that was supposed to listen to the clearing of all filters through checking weather or not there are any fitlerFns will just reset the state immediately. Bug.!

With this commit the issue is resolved, no more bad practice and less bugs to deal with...