feat(web): Add unique validation; Vehicle regTags
Code was refactored to handle error client side. This was done
for ease of use. A custom hook was created that uses UseMemo
to cache expensive calculations and will only recaluclate when
regTags is changed. Data is fetched from the backend (useVehiclesQ)
and the regTag of all vehicles if validated. All of this is converted
into a constant so that it can be used as a ZodSchema.