feat(web): Persisting Portal Filters
Filters for the Portal were persisted. This is done so that
users of the portal don't have to get annoyed when they switch tabs or
which section they are working in. This was done by changing from
useState to usePersistentState + useSlugs (that was edited). Within
FilterSearch onChange was simplified and the bit of code within it was
moved to a useEffect outside it for a proper persistent search bar.