fix(web): Cannot Edit Vehicles; regTag Unique applies in edit mode
The current regTag gets excluded from the the unqiue regTag validation.
When editing a vehicle it wouldn't work as the unqiue regTag validation
would get upset that the regTag of the vehicle that is being edited is
already being used. The current regTag gets excluded based on the id
parameter from the url, so if the reg is like the regTag from the
current url it gets excluded.