fix(web): DataTable functions not memoise causes infinite renders in useEffects
Before we have not used functions in useEffect dependency arrays but since it is best practice we should allow the possability. There we wrap all the functions from useHeadlessTable (used by useDataTable) in useCallback and also one in the popover filter

Note that we use the setState(x => newX) pattern to minimise the dependencies for the useCallback.