fix(web): "See Tickets" vehicle action not working.
Init state previously overode the value if the value from the store was falsy. This led to some problems though, multi select filter have init state as an empty array and thus, as soon as that empty array is persisted any new init state passed to usePersistentState will never be set.

Enter the possibility to dynamically set the initState using a function.