fix(web): Persisted filters can prevent expected visible tickets
An interesting issue with persisted filter states. As "See Tickets" shortcut will preselect that vehicle in the vehicles filter on the tickets page. If other persisted filters are selected that will not allow tickets that have the vehicle to be seen the user will end up on the no filtered results table. Uh oh.

This solution is quite hefty, one might argue that it is better to remove the shortcut all together because of the increased complexity, but complication can be fun and good UX is like gold to the developer soul.

We solve the issue by clearing any filterers that conflict the vehicle selection. Thus keeping persisted filters from the user if the work with the shortcut, those that dont disappear.

Navigating to a ticket, either by history or if pressing on a ticket in the vehicle table should also remove conflicting filters