feat(web): Date filter for audit columns
As we add createdAt and updatedAt columsn to all pages the date filter recieves some upgrades.

It is packages in a PortalDateFilter to contain more than just the calendar; a select dropdown to choose which oclumn you want to filter on.

The date filter is also upgraded to take a disabled prop as well as some more effects to handle date ranges that are out of the accepted data set.

Plus a fix on persistent state which lacked a key.