refactor(web): Auth api file upgrades
Rewrite the return statement of the interceptor to set the currentRefreshrequest and invoke the promise as its value in line. Not that i think that the promise could complete before return but this way it cannot and reduces the steps to follow.

Also inludes "fix" for unhapy typescript regarding accessing the process variable for settin glogin and log out on the window object