refactor(api): Cannot Create Edit or Delete Users on Admin/Users
Changed code to enable editing and deletion of users, even more
code was changed to do the same for creation of users.

Creating, editing and deleting users lead to errors, after first
fix problems with creating remained. Hard coded ID in data.sql
lead to the portal attempting to create Users with IDs already
in use.

UserResponse was made unrequired for the more reliable UserDto.
GenerationType for ID in User.java is now IDENTITY, removed the
ID parameter in data.sql for users to be more identitcal to other
model classes.

Solves issue #155