fix: Unable to create new employees in Admin portal
Fixes the issue preventing new employees from being created in the admin portal, removes the `employeeNumber` property from both the backend and frontend to align with the current data structure. Employee properties in the frontend now correctly match the backend. It makes the personal number and comment fields optional for both customers and employees. Backend validation for personal numbers now matches the frontend implementation. It also fixes a bug where fields that are edited and then cleared are sent as empty strings instead of `undefined`. Additionally, it allows personal number /age field to be missing for both customer and employee.

Solves #143