chore: Unify approach to env vars
This way all environment variables needed to run/build the backend and frontend is stored in the same place.

The .idea directory is updated tso that the run configuration for spring read from the root .env file and s3 related config is fixed to read from environment variables, but thorugh application.properties instead which i believe is a good practice.

The previous env.properties file was some kind of hybrid of a java .properties file and environment variables but the intent was always for it to be env variables and not tied to java.

This means that when the aws ec2 instance runs the jar file it just has to populate these environment variables to bash and run the jar and it works.

This also goes for say the pipeline building the frontend, it will have the environment variables ready and therefore will be able to build and make the requests needed to the backend....

All config needed to build and run and do local develpoment comes from the same source.

In some cases and operative systems (windows!) will need the .env file to work wile sometimes just having the .env vars read will work as well.

The ec2 instance will probably use docker-compose anyway and that has native support for .env file.