fix: Lack of env-var support IDEA Maven prevent build
While the previous commit resolved a unified solution for environment variables in production (running the .jar) and when using the spring boot run configuration (IDEA Ultimate) when running maven in IDEA it does not access environment variables from bash or of course form any .env file.

This solution uses the dotenv dependency and a new initialiser that consumes it to load environment variables from an .env file if present in the root or in the project directory (as of now main/).

This way it works when running the .jar file thourgh native spring support of env vars and when developing/building locally, regardless of IDEA Ultimate or not.

More explenation available in the context initialiser class.