refactor(api): Refactored User code for ease of reading
Code surrounding user was convoluted and hard to maintain. If
new fields were ever added to User, then changes to the code
would have had to be made in numerous locations. To fix this
and to make the code more familiar it was changed to resemble
other sections of the code.