feat(web): Merge csr & ssr routes and localise all routes
The idea to have csr route being non localised and ssr routes being localised routes is not sustainable.

Portal is a csr route, and vehicle details will be one as well. But vehicle detaisl will need to be localised so the tight coupling between csr/ssr and localisation is not correct.

After some thoguht we decided to just localise all routes instead. Why not.