refactor: New applyToEntity design pattern + localisation service
The localisation utils wasn't entierly ready, we had not recognised the need to object mapper and working with json for localised rich text values.

Creating a localisation service instead works better. This also opened up some reflections on the current design pattern, it is more explicit if a dtos applyToEntity method actaully takes everything it needs as arguments. This relies less on remembering to sett the field in the service and more on providing the needed spec.

README updated to reflect the new, somewhat more refined approach