feat: Implement web edit in the portal
Becuase of unfortunate insidednts and bad planning a few things have coincided into this big commit.

First of all the three pages, text, shops and stories are now editable in the portal.

The abstraction of language based editing give a new nuance to the flow and usage of the form.

Rich text editing is also added.

Styles for translations are added and touches on various palces – with this some general design is also retouched lightly.

The addition of navbar using icons (and tooltips for readabiliy) is also included and elevates the experience somewthat....