feat: Implement web edit in the portal
Because of unfortunate incidents and bad planning a few things have coincided into this big commit.

First of all the three pages, text, shops and stories are now editable in the portal.

The abstraction of language based editing give a new nuance to the flow and usage of the form.

Rich text editing is also added.

Styles for translations are added and touches on various places – with this some general design is also retouched lightly.

The addition of navbar using icons (and tooltips for readability) is also included and elevates the experience somewhat....

Also – the edition of actual enums for Text is a big improvement, resulting in a more actual contract. This means that we now know the fixed amount of text we have, and as icing on the cake get typescript inference when working with text from usePrelaodedDataLocalise()...

The dev db will need some cleaning up as I've with the reconciliation of texts transitioned unnecessary text to the code based localization and left longer sentences only...